+++ ui/packages/shared/profile/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.16.395 (2024-07-01)
+
+**Note:** Version bump only for package @parca/profile
+
 ## [0.16.394](https://github.com/parca-dev/parca/compare/@parca/profile@0.16.393...@parca/profile@0.16.394) (2024-06-27)
 
 **Note:** Version bump only for package @parca/profile